Add support for parameterized runner at method level
The JUnit Parameterized runner allows to run many times a test methods
with different parameters. From the Android command line tool,
`am instrumentation`, without the output in raw format (using `-r`
option), this is the same than a normal test method, but
the outcome test results are understood as:
- ignored when all are ignored
- failure when one is failing
- success when all are successful or some are ignored (*)

In case of success, the string "OK (XX tests)" is printed instead of
"OK (1 test)".

Note that to have a finer granularity and fully handle the case (*),
then the raw format should be supported first.